meta-openembedded: subtree update:7c02c7d411..713e70bd25
Adam Miartus (1):
      toybox: bump version to 0.8.3

Adrian Bunk (7):
      libhandy: Remove, moved to OE-core
      unicode-ucd: Stop broken license downloading
      python3-docutils: Remove, moved to OE-core
      python3-dbusmock: Remove, moved to OE-core
      python3-cython: Remove, moved to OE-core
      python3-pyparsing: Remove, moved to OE-core
      postfix: Upgrade 3.4.10 -> 3.4.12

Alistair Francis (1):
      openocd: Update libjaylink

Andreas Müller (4):
      libusb-compat: upgrade 0.1.5 -> 0.1.7
      hidapi: upgrade 0.7.99+ -> 0.9.0
      gexiv2: upgrade 0.12.0 -> 0.12.1
      thunar: upgrade 1.8.14 -> 1.8.15

Andrew Geissler (1):
      nlohmann-json: backport gcc10 fix

Armin Kuster (1):
      tremor: update SRC_URI as project moved to gitlab

Beniamin Sandu (1):
      rasdaemon: create recipe for version 0.6.5

Changqing Li (1):
      apache2: fix service start fail

Christoph Steiger (1):
      python-periphery: Add python-mmap to RDEPENDS

Denys Dmytriyenko (2):
      python3-pycryptodome(x): moved to OE-Core, remove from meta-python
      python3-pyelftools: moved to OE-Core, remove from meta-python

Domarys Correa (18):
      python3-argh: Add recipe version 0.26.2
      python3-graphviz: Add recipe version 0.14
      python3-aiohttp-jinja2: Add recipe version 1.2.0
      python3-py-ubjson: Add recipe version 0.16.1
      python3-trafaret: Add recipe version 2.0.2
      python3-trafaret-config: Add recipe version 2.0.2
      python3-txaio: Add recipe version 20.4.1
      python3-u-msgpack-python: Add recipe version 2.6.0
      python3-xmodem: Add recipe version 0.4.6
      python3-multidict: Update 4.7.4 -> 4.7.5
      python3-lz4: Add recipe version 3.0.2
      python3-autobahn: Add recipe version 20.4.3
      python3-snappy: Add recipe version 0.5.4
      python3-cbor2: Add recipe version 5.1.0
      python3-pathtools3: Add recipe version 0.2.1
      python3-watchdog: Add recipe version 0.10.2
      python3-jinja2: Update 2.11.1 -> 2.11.2
      python3-ansicolors: Add recipe version 1.1.8

Emmanuel Roullit (1):
      appstream-glib: update 0.7.16 -> 0.7.17

Honggyu Kim (1):
      uftrace: add PV to update the package version

Kai Kang (4):
      garcon: 0.6.4 -> 0.7.0
      xfconf: 4.14.2 -> 4.14.3
      thunar: 1.8.12 -> 1.8.14
      catfish: 1.4.11 -> 1.4.13

Khem Raj (5):
      gtkwave: Disable mime update during install
      netplan: Depend on systemd if it is in distro
      postfix: Upgrade to 3.4.10 and compile with -fcommon
      uim: Add patch to fix -fno-common link error
      safec: Update to latest on 3.5.1 release tags

Leon Anavi (17):
      transmission: Add version 2.94
      pamela: Upgrade 0.3.0 -> 1.0.0
      python3-absl: Upgrade to 0.9.0
      python3-gmqtt: Upgrade to 0.6.5
      python3-appdirs: Upgrade to 1.4.4
      python3-aiofiles: Upgrade to 0.5.0
      python3-markdown: Upgrade to 3.2.2
      python3-luma-core: Upgrade to 1.14.0
      python3-pandas: Upgrade 1.0.1 -> 1.0.3
      python3-parallax: Upgrade 1.0.5 -> 1.0.6
      python3-openpyxl: Upgrade 2.6.3 -> 3.0.3
      python3-setuptools-scm: Upgrade 3.3.3 -> 3.5.0
      python3-toml: Extend with native build variants
      python3-zipp: Upgrade 0.6.0 -> 3.1.0
      python3-ujson: Upgrade 1.35 -> 2.0.3
      python3-webcolors: Upgrade 1.8.1 -> 1.11.1
      python3-pid: Upgrade 2.2.5 -> 3.0.3

Michael Haener (2):
      cockpit: upgrade 218 -> 219
      cockpit: rt-deps for storaged

Paul Barker (1):
      mplayer-common: Drop obsolete recipe

Pierre-Jean Texier (4):
      iwd: upgrade 1.6 -> 1.7
      librsync: upgrade 2.3.0 -> 2.3.1
      ser2net: fix upstream check URL
      ser2net: upgrade 4.1.5 -> 4.1.8

Robert P. J. Day (1):
      meta-python: delete superfluous python-mako.inc

Sean Nyekjaer (2):
      gpsd: fix build on aarch64
      gpsd: mark CLEANBROKEN

Trevor Gamblin (2):
      meta-python: add self as maintainer
      python3-iso8601: add python3-numbers to RDEPENDS

Wang Mingyu (6):
      modemmanager: upgrade 1.12.8 -> 1.12.10
      python3-pyroute2: upgrade 0.5.11 -> 0.5.12
      jansson: upgrade 2.12 -> 2.13.1
      openldap: upgrade 2.4.49 -> 2.4.50
      python3-pycparser: upgrade 2.19 -> 2.20
      cryptsetup: upgrade 2.3.1 -> 2.3.2

William A. Kennington III (1):
      cli11: upgrade 1.8.0 -> 1.9.0

Yanfei Xu (1):
      turbostat: fix the build failure for new v5.7-rc6 kernel

Yi Zhao (3):
      libldb: upgrade 1.5.6 -> 1.5.7
      samba: upgrade 4.10.13 -> 4.10.15
      apache2: create log/run directory via pkg_postinst

Zang Ruochen (2):
      dnsmasq: upgrade 2.80 -> 2.81
      fetchmail: upgrade 6.4.3 -> 6.4.4

Zheng Ruoqin (3):
      dstat: Fix runtime error that depend python.
      kea: upgrade 1.7.6 -> 1.7.7
      libqmi: upgrade 1.24.8 -> 1.24.12

+++ meta-openembedded/meta-oe/recipes-kernel/turbostat/turbostat_3.4.bb
@@ -42,16 +42,28 @@ do_configure_prepend() {
 	mkdir -p ${S}
 	cp -r ${STAGING_KERNEL_DIR}/arch/x86/include/asm/msr-index.h ${S}
 	cp -r ${STAGING_KERNEL_DIR}/arch/x86/include/asm/intel-family.h ${S}
-	cp -r ${STAGING_KERNEL_DIR}/include/linux/bits.h ${S}
-	cp -r ${STAGING_KERNEL_DIR}/include/linux/const.h ${S}
+	if [ -f "${STAGING_KERNEL_DIR}/include/vdso/bits.h" ]; then
+		cp -r ${STAGING_KERNEL_DIR}/include/vdso/bits.h ${S}
+		cp -r ${STAGING_KERNEL_DIR}/include/vdso/const.h ${S}
+	else
+		cp -r ${STAGING_KERNEL_DIR}/include/linux/bits.h ${S}
+		cp -r ${STAGING_KERNEL_DIR}/include/linux/const.h ${S}
+	fi
 	cp -r ${STAGING_KERNEL_DIR}/tools/power/x86/turbostat/* ${S}
 	cp -r ${WORKDIR}/COPYING ${S}
 }
 
+
 do_compile() {
 	sed -i 's#<linux/bits.h>#"bits.h"#' msr-index.h
-	sed -i 's#<linux/const.h>#"const.h"#' bits.h
-	sed -i -e 's#<uapi/linux/const.h>#<linux/const.h>#' -e 's#_LINUX_CONST_H#_LINUX_CONST_H_KERNEL#' const.h
+	'TMPCHECK='grep "<vdso/const.h>" bits.h'' || true
+	if [ -n $TMPCHECK ]; then
+		sed -i 's#<vdso/const.h>#"const.h"#' bits.h
+		sed -i 's#<uapi/linux/const.h>#<linux/const.h>#' const.h
+	else
+		sed -i 's#<linux/const.h>#"const.h"#' bits.h
+		sed -i -e 's#<uapi/linux/const.h>#<linux/const.h>#' -e 's#_LINUX_CONST_H#_LINUX_CONST_H_KERNEL#' const.h
+	fi
 	sed -i 's#MSRHEADER#"msr-index.h"#' turbostat.c
 	sed -i 's#INTEL_FAMILY_HEADER#"intel-family.h"#' turbostat.c
 	sed -i 's#\$(CC) \$(CFLAGS) \$< -o \$(BUILD_OUTPUT)/\$@#\$(CC) \$(CFLAGS) \$(LDFLAGS) \$< -o \$(BUILD_OUTPUT)/\$@#' Makefile
